+++ assembly/contracts/serializable/Ticket.ts
@@ -4,9 +4,9 @@ import { JSON } from 'json-as';
 export class Ticket {
   address: string;
   numbers: u8[];
-  no: u8;
+  no: u64;
 
-  constructor(address: string, numbers: u8[], no: u8) {
+  constructor(address: string, numbers: u8[], no: u64) {
     this.address = address;
     this.numbers = numbers;
     this.no = no;
